세이브 & 로드 기능 완성

+++ OpenRPG/Tiles/Tile.h
@@ -12,7 +12,7 @@ protected:
 
 public:
 	Tile();
-	Tile(float x, float y, float gridSizeF, const sf::Texture& texture, const sf::IntRect& texture_rect,
+	Tile(unsigned x, unsigned y, float gridSizeF, const sf::Texture& texture, const sf::IntRect& texture_rect,
 		bool collision = false, short type = TileTypes::DEFAILT);
 	virtual ~Tile();
